Replace arrow caret in segmenter card with gear icon
